+++ apps/hubble-stats/containers/charts/OverviewTvlChartContainer.tsx
@@ -1,11 +1,11 @@
-import getTvlChartData from '../../data/getTvlChartData';
+import getOverviewTvlChartData from '../../data/charts/getOverviewTvlChartData';
 import { AreaChartContainerClient } from './private';
 import { ChartProps } from './types';
 
 export default async function OverviewTvlChartContainer(props: ChartProps) {
   const { numDatesFromStart, startingEpoch } = props;
 
-  const { tvlData, currentTvl } = await getTvlChartData(
+  const { tvlData, currentTvl } = await getOverviewTvlChartData(
     startingEpoch,
     numDatesFromStart
   );
