+++ apps/hubble-stats/containers/charts/OverviewVolumeChartContainer.tsx
@@ -1,11 +1,11 @@
-import getVolumeChartData from '../../data/getVolumeChartData';
+import getOverviewVolumeChartData from '../../data/charts/getOverviewVolumeChartData';
 import { VolumeChartContainerClient } from './private';
 import { ChartProps } from './types';
 
 export default async function OverviewVolumeChartContainer(props: ChartProps) {
   const { numDatesFromStart, startingEpoch } = props;
 
-  const { volumeData, deposit24h } = await getVolumeChartData(
+  const { volumeData, deposit24h } = await getOverviewVolumeChartData(
     startingEpoch,
     numDatesFromStart
   );
